Issue #10981: Moved GroupEntityDeriverTest to drupal-test-traits
